% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{home_range_cpp}
\alias{home_range_cpp}
\title{Main home range estimation function}
\arguments{
\item{data}{data cube of the state space}

\item{par}{parameters for model execution or fitting}

\item{trajectoryPath}{location of the track file}

\item{resolution}{resolution (dimensions) of the state space (as not georeferenced)}

\item{nSimulatedSteps}{number of simulated steps to run}

\item{nSimulatedRuns}{number of runs to simulate}

\item{optimization}{return optimization outputs (TRUE or FALSE)}

\item{verbose}{provide verbose output (TRUE or FALSE)}
}
\value{
simulated model locations or a target value for optimization
(cost function)
}
\description{
C++ back-end function doing all the heavy lifting
}
