% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr_convert_drivers.R
\name{hr_convert_drivers}
\alias{hr_convert_drivers}
\title{Converts data to a consistent format}
\usage{
hr_convert_drivers(r, par, na_fill = 0)
}
\arguments{
\item{r}{a terra SpatRaster}

\item{par}{a parameter file}

\item{na_fill}{fill value for NA values (default = 0)}
}
\value{
a nested list of data to be used in model prediction and fitting
}
\description{
Data conversion tool to generate a consistent
input format specific to the package
}
