% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr_cost.R
\name{hr_cost}
\alias{hr_cost}
\title{Optimization cost function}
\usage{
hr_cost(par, obs, data, names)
}
\arguments{
\item{par}{parameters}

\item{obs}{reference observations}

\item{data}{driver matrix}

\item{names}{names of the parameters}
}
\value{
log likelihood values given a data and parameter set
}
\description{
Wrapper around the low level Cpp function, avoids some of the
sanity checks of hr_run() thereby saving time during optimization.
}
\details{
The input assumes regularized data using hr_regularize() for both
the spatial grid and the observed tracks.
}
