% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr_dl_maps.R
\name{hr_dl_maps}
\alias{hr_dl_maps}
\title{Download raster maps}
\usage{
hr_dl_maps(
  track,
  settings = data.frame(collection = c("esa-worldcover", "cop-dem-glo-30"), year =
    c(2021, 2021), asset = c("map", "data"), filename = c("LC.tif", "DEM.tif")),
  buffer = 20,
  path = tempdir(),
  overwrite = TRUE,
  progress = FALSE
)
}
\arguments{
\item{track}{bio-logging track with coordinates in lat lon (sf object)}

\item{settings}{downloads settings, data frame specifying data
products and their target year and assets on the planetary computer STAC}

\item{buffer}{buffer (in km, default = 20) around the bio-logging track file
to accommodate for a sufficiently large home range area. Value should be
adjusted to the target species.}

\item{path}{output path}

\item{overwrite}{overwrite / and re-download the data}

\item{progress}{show download progress bar (as feedback on large downloads,
default = FALSE)}
}
\value{
raster maps covering your area of interest as geotiff
}
\description{
This downloads the basic raster maps covering your track, including a
buffer zone to account for ranging behaviour. Note that this does not
take care of data conversions or data harmoization.
}
\details{
This function does not leverage cloud optimized geotiffs via GDAL's
vsicurl options due to the inconsistency in the deployment across the STAC.
Therefore tiled data is downloaded which comes at a slight data volume
penalty in favour of the stability across data products.

Please consult the Planetary computer data repository for information on the
STAC product collection name, the required year and the asset (subset)
required.
}
