% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr_drivers.R
\name{hr_drivers}
\alias{hr_drivers}
\title{Create spatial driver data for model input}
\usage{
hr_drivers(
  track,
  crs = "EPSG:3035",
  buffer = 7,
  window = 11,
  path,
  na_value = NA,
  overwrite = TRUE
)
}
\arguments{
\item{track}{observed locations track}

\item{crs}{an equal area CRS projection reference (default = EPSG:3035,
or extended LAEA Europe)}

\item{buffer}{buffer (in km, default = 11) around the bio-logging track file
to accommodate for a sufficiently large home range area. Value should be
adjusted to the target species.}

\item{window}{window size for resampled forest cover values, number of pixels
of the window (uneven value, default = 11 or approx. 300m)}

\item{path}{path and filename where to save the raster map data. When using
a temporary file use file.path(tempdir(), "raster_drivers.tif") or similar}

\item{na_value}{value with which to substitute NA values (e.g. -9999)}

\item{overwrite}{overwrite output if the path name is the same (default = TRUE)}
}
\value{
nested list with the regularized data and observed track, also saves
the used projection crs
}
\description{
Regularizes spatial input data to a common equal area grid. It also performs
sanity checks on the inputs to check if observations cover the spatial grid.
Data can be provided as arguments, or if no inputs are provided the data will
be downloaded in the background.
}
\details{
It is encouraged to save the data to file, rather than use a returned object.
This limits the memory footprint of operations.
}
