% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr_format_drivers.R
\name{hr_format_drivers}
\alias{hr_format_drivers}
\title{Format drivers}
\usage{
hr_format_drivers(track, id, secret)
}
\arguments{
\item{track}{biologging track with coordinates in lat lon (sf object)}

\item{id}{planetary computer id}

\item{secret}{planetary computer secret}
}
\value{
a nested list with the data cube or state space to use, the converted
X/Y coordinates relative to the data cube, and the geo-referencing information
for mapping and back converting the estimated locations
}
\description{
Download and convert drivers for a given track.
}
