% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr_predict.R
\name{hr_predict}
\alias{hr_predict}
\title{Predict home range model values}
\usage{
hr_predict(
  data,
  par,
  obs,
  steps = 1,
  runs = 1,
  keep_start = TRUE,
  optimization = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{data}{A data frame with driver data.}

\item{par}{model parameters}

\item{obs}{reference observations (path) to a track file generated by \{hr_xy()\}}

\item{steps}{number of steps to simulate}

\item{runs}{number of runs to simulate}

\item{keep_start}{keep the starting position in the output data (this is
run 0 and only has one coordinate, the release point)}

\item{optimization}{optimization trigger (only used in debugging, set to FALSE)}

\item{verbose}{provide feedback on the command line (default = TRUE)}
}
\value{
simulated track
}
\description{
This is the main function to predict values from a fit set of
parameters.
}
