% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr_regularize.R
\name{hr_regularize}
\alias{hr_regularize}
\title{Regularize spatial input data}
\usage{
hr_regularize(data, obs, crs = "EPSG:3035")
}
\arguments{
\item{data}{raster stack with geospatial input data}

\item{obs}{observed locations track}

\item{crs}{an equal area CRS projection reference (default = EPSG:3035,
or extended LAEA Europe)}
}
\value{
nested list with the regularized data and observed track, also saves
the used projection crs
}
\description{
Regularizes spatial input data to a common equal area grid. It also performs
sanity checks on the inputs to check if observations cover the spatial grid
and returns the
}
