% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr_xy.R
\name{hr_xy}
\alias{hr_xy}
\title{Convert map / track records to col/row (X/Y) coordinates}
\usage{
hr_xy(map, track, method = "exact")
}
\arguments{
\item{map}{a geo-referenced \{terra\} raster map (stack)}

\item{track}{an \{sf\} multi-point file of observations}

\item{method}{exact or approximate location conversions (default = "exact")}
}
\value{
the location of the track file in the correct format (written to
the temporary directory)
}
\description{
Converts map and track coordinates to col/row (X/Y), relative values with
respect to the provided map. This is an internal function to interface
with the C++ code which is not georeferenced internally. Note that you will
need to use an equal area projection for best performance. For small areas
lon/lat (WGS84) can be used but is not encouraged. The project should be
chosen based on the locale.
}
