% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{par_delist}
\alias{par_delist}
\title{De-list parameters}
\usage{
par_delist(par)
}
\arguments{
\item{par}{home range model parameter settings}
}
\value{
long format parameter file
}
\description{
Converts nested parameter lists to a long format
where the vector parameter 'coef' is split in multiple
parts. This allows for BayesianTools priors to be generated
from the nested parameter files used in the model call, and
allows for flexible use of driver files.
}
