% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.hr_predict}
\alias{plot.hr_predict}
\title{Plotting method for prediction results}
\usage{
\method{plot}{hr_predict}(x, ...)
}
\arguments{
\item{x}{input data generated using the pr_fit() function}

\item{...}{additional parameters to pass}
}
\value{
a plot (map) of predicted statistics
}
\description{
Plotting method for prediction results
}
\keyword{behaviour}
\keyword{model,}
