% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{z_score}
\alias{z_score}
\title{z-score}
\usage{
z_score(r)
}
\arguments{
\item{r}{terra raster with a single layer}
}
\value{
a z-score normalized map using the global mean and sd
}
\description{
Scale and center the data of a continuous map using a z-score.
}
