% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr_cost.R
\name{hr_cost}
\alias{hr_cost}
\title{Optimization cost function}
\usage{
hr_cost(par, obs, data, resolution, names)
}
\arguments{
\item{par}{A list containing model calibration settings}

\item{obs}{reference observations, a track file generated by \{hr_xy()\}}

\item{data}{input data array}

\item{resolution}{resolution of the underlying driver data}

\item{names}{names of the parameters}
}
\value{
log-likelihood values given a data and parameter set
}
\description{
Low level wrapper around the low level Cpp function for optimization
purposes
}
\details{
The input assumes regularized data using hr_regularize() for both
the spatial grid and the observed tracks.
}
