% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr_fit.R
\name{hr_fit}
\alias{hr_fit}
\title{Fit a home range model}
\usage{
hr_fit(data, obs, par, resolution, optim_out = TRUE, parallel = FALSE, ...)
}
\arguments{
\item{data}{A data frame with driver data.}

\item{obs}{reference observations (path) to a track file generated by \{hr_xy()\}}

\item{par}{A list containing model calibration settings.}

\item{resolution}{resolution of the underlying maps (grid)}

\item{optim_out}{A logical indicating whether the function returns the raw
output of the optimization functions (defaults to TRUE).}

\item{parallel}{support parallel processing}

\item{...}{Optional arguments passed on to the cost function specified as
\code{par$metric}.}
}
\value{
A named list containing the calibrated parameter vector \code{par} and
the output object from the optimization \code{mod}. For more details on this
output and how to evaluate it, see \link[BayesianTools:runMCMC]{runMCMC} (also
\href{https://florianhartig.github.io/BayesianTools/articles/BayesianTools.html}{this post})
}
\description{
This is the main function that handles the
calibration of the home_range_cpp() C++ function
through hr_cost().
}
