% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr_fit.R
\name{hr_fit}
\alias{hr_fit}
\title{Fit a home range model}
\usage{
hr_fit(drivers, obs, settings, optim_out = TRUE, ...)
}
\arguments{
\item{drivers}{A data frame with driver data. See \code{\link{p_model_drivers}}
for a description of the data structure.}

\item{obs}{A data frame containing observational data used for model
calibration. See \code{\link{p_model_validation}} for a description of the data
structure.}

\item{settings}{A list containing model calibration settings.}

\item{optim_out}{A logical indicating whether the function returns the raw
output of the optimization functions (defaults to TRUE).}

\item{...}{Optional arguments passed on to the cost function specified as
\code{settings$metric}.
.}
}
\value{
A named list containing the calibrated parameter vector \code{par} and
the output object from the optimization \code{mod}. For more details on this
output and how to evaluate it, see \link[BayesianTools:runMCMC]{runMCMC} (also
\href{https://florianhartig.github.io/BayesianTools/articles/BayesianTools.html}{this post})
and \link[GenSA]{GenSA}.
}
\description{
This is the main function that handles the
calibration of the home_range_cpp() C++ function
through hr_cost().
}
