% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr_predict.R
\name{hr_predict}
\alias{hr_predict}
\title{Title}
\usage{
hr_predict(
  data,
  par,
  obs,
  resolution,
  steps = 1,
  runs = 1,
  optimization = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{data}{data}

\item{par}{parameters}

\item{obs}{observed track}

\item{resolution}{resolution}

\item{steps}{steps}

\item{runs}{runs}

\item{optimization}{optimize}

\item{verbose}{feedback}
}
\value{
simulated track
}
\description{
Title
}
