% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr_regularize.R
\name{hr_regularize}
\alias{hr_regularize}
\title{Regularize spatial input data}
\usage{
hr_regularize(dem, lc, track, crs = "EPSG:3035")
}
\arguments{
\item{dem}{DEM terra raster}

\item{lc}{land cover map (ESA world cover)}

\item{track}{observed locations track}

\item{crs}{an equal area CRS projection reference (default = EPSG:3035,
or extended LAEA Europe)}
}
\value{
nested list with the regularized data and observed track, also saves
the used projection crs
}
\description{
Regularizes spatial input data to a common equal area grid. It also performs
sanity checks on the inputs to check if observations cover the spatial grid.
Data can be provided as arguments, or if no inputs are provided the data will
be downloaded in the background.
}
